+++ src/plcc/code/class_.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Class:
+    ...
