+++ src/plcc/code/module.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Module:
+    ...
