+++ src/plcc/code/names.py
@@ -5,16 +5,6 @@ from plcc.spec.bnfrule import Symbol
 
 
 @dataclass
-class File:
-    ...
-
-
-@dataclass
-class Class:
-    ...
-
-
-@dataclass
 class UnresolvedTypeName:
     symbol: Symbol
