+++ src/plcc/spec/parser/bnfparser.py
@@ -1,8 +1,8 @@
 import re
 
-from ..spec.bnfrule import BnfRule
-from ..spec.bnfrule import Symbol
-from ..spec.bnfspec import BnfSpec
+from ..bnfrule import BnfRule
+from ..bnfrule import Symbol
+from ..bnfspec import BnfSpec
 
 
 class BnfParser:
