+++ src/plcc/spec/validator/lexvalidator.py
