+++ tests/unit/test_astdesigner.py
@@ -1,13 +1,13 @@
 import pytest
 
 
-from plcc.designer.ast import AstDesigner
+from plcc.code.generator.ast import AstDesigner
 from plcc.spec.bnfspec import BnfSpec
 from plcc.spec.bnfrule import BnfRule
 from plcc.spec.symbol import Symbol
-from plcc.parser.bnfparser import BnfParser
-from plcc.parser.specreader import SpecReader
-from plcc.translator.code import File
+from plcc.spec.parser.bnfparser import BnfParser
+from plcc.spec.parser.specreader import SpecReader
+from plcc.code.module import Module
 
 
 @pytest.fixture
@@ -33,5 +33,5 @@ def test_single_empty_bnf_rule_returns_one_class(ast_designer):
     spec = givenBnfSpec('<one> ::=')
     astClassFiles = ast_designer.design(spec)
     assert len(astClassFiles) == 1
-    assert isinstance(astClassFiles[0], File)
+    assert isinstance(astClassFiles[0], Module)
